feat(useexpander): extract useAccordion to be its own hook
